requirements.txt
transformers
torch
google-genai
gradio
langgraph
langchain
langchain-core
langchain_google_genai
pytesseract
pillow
dotenv
langgraph-checkpoint-sqlite
scikit-learn
